+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## [master](https://github.com/c-sp/gameboy-test-roms)
 
 - use
+  [mealybug-tearoom-tests:70e88fb90b](https://github.com/mattcurrie/mealybug-tearoom-tests/tree/70e88fb90b)
+  *(2020-12-19)*
+- use
   [age-test-roms:fb6cd3b1d7](https://github.com/c-sp/age-test-roms/tree/fb6cd3b1d7)
+  *(2021-05-28)*
 - use
   [cgb-acid-hell:107b7c5a87](https://github.com/mattcurrie/cgb-acid-hell/tree/107b7c5a87)
   *(2021-03-22)*